+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Buttons/Krypton.Toolkit.Suite.Extended.Buttons 2022.csproj
@@ -78,7 +78,7 @@
 					</PropertyGroup>
 
 					<ItemGroup>
-						<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.63-alpha" />
+						<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.67-alpha" />
 					</ItemGroup>
 				</When>
