+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Core/Krypton.Toolkit.Suite.Extended.Core 2022.csproj
@@ -81,15 +81,15 @@
 		  </PropertyGroup>
 
 		  <ItemGroup>
-			  <PackageReference Include="Krypton.Docking.Nightly" Version="65.22.2.63-alpha" />
+			  <PackageReference Include="Krypton.Docking.Nightly" Version="65.22.2.67-alpha" />
 			  
-			  <PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.63-alpha" />
+			  <PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.67-alpha" />
 
-			  <PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.63-alpha" />
+			  <PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.67-alpha" />
 
-			  <PackageReference Include="Krypton.Ribbon.Nightly" Version="65.22.2.63-alpha" />
+			  <PackageReference Include="Krypton.Ribbon.Nightly" Version="65.22.2.67-alpha" />
 
-			  <PackageReference Include="Krypton.Navigator.Nightly" Version="65.22.2.63-alpha" />
+			  <PackageReference Include="Krypton.Navigator.Nightly" Version="65.22.2.67-alpha" />
 		  </ItemGroup>
 	  </When>
