+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Developer.Utilities/Krypton.Toolkit.Suite.Extended.Developer.Utilities 2022.csproj
@@ -75,7 +75,7 @@
 		  </PropertyGroup>
 
 		  <ItemGroup>
-			  <PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.63-alpha" />
+			  <PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.67-alpha" />
 		  </ItemGroup>
 	  </When>
