+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Error.Reporting/Krypton.Toolkit.Suite.Extended.Error.Reporting 2022.csproj
@@ -77,8 +77,8 @@
 			</PropertyGroup>
 
 			<ItemGroup>
-				<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.63-alpha" />
-				<PackageReference Include="Krypton.Navigator.Nightly" Version="65.22.2.63-alpha" />
+				<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.67-alpha" />
+				<PackageReference Include="Krypton.Navigator.Nightly" Version="65.22.2.67-alpha" />
 			</ItemGroup>
 		</When>
