+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Outlook.Grid/Krypton.Toolkit.Suite.Extended.Outlook.Grid 2022.csproj
@@ -76,7 +76,7 @@
 			</PropertyGroup>
 
 			<ItemGroup>
-				<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.63-alpha" />
+				<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.67-alpha" />
 			</ItemGroup>
 		</When>
