+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Theme.Switcher/Krypton.Toolkit.Suite.Extended.Theme.Switcher 2022.csproj
@@ -78,9 +78,9 @@
 			</PropertyGroup>
 
 			<ItemGroup>
-				<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.63-alpha" />
+				<PackageReference Include="Krypton.Toolkit.Nightly" Version="65.22.2.67-alpha" />
 
-				<PackageReference Include="Krypton.Navigator.Nightly" Version="65.22.2.63-alpha" />
+				<PackageReference Include="Krypton.Navigator.Nightly" Version="65.22.2.67-alpha" />
 			</ItemGroup>
 		</When>
